Add SecurityPermissionAttribute to netstandard ref
